// libs/api/src/dto/src/invoice-item.dto.ts
import { CountingUnitDto } from "./counting-unit.dto";
import { GoodsDto } from "./goods.dto";

export interface InvoiceItemDto {
  id?: number;
  goodsService?: GoodsDto;
  quantity?: number;
  fee?: number;
  discountAmount?: number;
  taxAmount?: number;
  additionsAmount?: number;
  deductionsAmount?: number;
  description?: string;
  measurement?: CountingUnitDto;
  isDeleted?: boolean;
}